requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
numpy>=1.21.0
pandas>=1.5.0
streamlit>=1.22.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0 